fix(drawer): header height for android
- extracted height functions to helper file
  - this is needed when merging with the surveys feature
- renamed calendarHelper file to be in line with other helper files
