+++ src/helpers/index.js
@@ -1,9 +1,10 @@
 export * from './addressHelper';
 export * from './bookmarkHelper';
-export * from './calendarHelpers';
+export * from './calendarHelper';
 export * from './dateTimeHelper';
 export * from './fileSizeHelper';
 export * from './graphqlHelper';
+export * from './headerHelper';
 export * from './htmlViewHelper';
 export * from './imageHelper';
 export * from './linkHelper';
